requirements.txt
google-generativeai==0.3.2
chromadb==0.4.18
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
pdfplumber==0.9.0
Pillow==10.1.0
gunicorn==21.2.0
